Tar ut enkel fnr til egen komponent under felleskomponenter
