+++ src/felleskomponenter/skjema/inputFnrDnrOrgnr/index.ts
@@ -1,0 +1,3 @@
+import EnkelFellesInputFnrDnrOrgnrSaksnr from "./EnkelFellesInputFnrDnrOrgnrSaksnr";
+
+export default EnkelFellesInputFnrDnrOrgnrSaksnr;
